requirements.txt
Django==1.10.7
djangorestframework==3.6.3
django-crequest==2016.3.16